+++ cell/test/u_cell_cfg_test.c
@@ -136,7 +136,7 @@
 
 /** Used for keepGoingCallback() timeout.
  */
-static int64_t gStopTimeMs;
+static int32_t gStopTimeMs;
 
 /** The GNSS profile bit map.
  */
